update forward_backward st for internvl2-1B

+++ align_interface_padding_mask/inference_internvl_v2.0_1B.py
@@ -18,7 +18,7 @@ print(f'mode: {mode}')
 
 path = "./pretrained/InternVL2-1B"
 
-model = InternVLChatModel.from_pretrained(path)
+model = InternVLChatModel.from_pretrained(path, ms_dtype=ms.float16)
 tokenizer = Qwen2Tokenizer.from_pretrained(path)
